requirements.txt
flask
numpy
pillow
python-dotenv
ultralytics
werkzeug
gevent